Co-locate autonomy candidate generators, rename module to autonomy.rs (-241)
Move the four autonomy-cascade candidate generators that remained in
main.rs (resident_waits_for_player_gift, resident_mutual_trade_candidate,
resident_gift_candidate, resident_delivery_candidate) into the module
that already owns the cascade, priority ladder, and rank constants.

Rename resident_offer_scoring.rs to autonomy.rs to reflect its actual
role: it is the home of the entire resident autonomy decision system
(cascade, inter-actor scheduler, rank tiers, record generation, ripple
candidate selection, intent projection, and now the candidate generators
themselves). The old name described a subset of the file's content.

Delete the dead RippleBudget::for_zone_and_action zone branch where the
ZONE_FRONTIER arm and the _ catch-all arm returned byte-identical
budgets. The zone is still tracked in RippleContext for downstream
filtering; only the budget computation was dead.

Behavior-preserving: 1184 tests pass, clippy at ceiling (225),
main.rs 63826 -> 63585. Version 1.0.62 -> 1.0.63.

Refs: autonomy extensibility Step 0 (co-locate the decision system).
