+++ v2/orchestrator-rust/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cosyworld-orchestrator"
-version = "1.0.62"
+version = "1.0.63"
 edition = "2021"
 publish = false
