+++ v2/scripts/main-rs-line-ceiling.txt
@@ -342,4 +342,9 @@
 # from main.rs into resident_offer_scoring.rs as pub(crate). The 3 AI-speech
 # collision methods (resident_reply_repeats_recent_event, collision_safe_*)
 # stay in main.rs. Pure, behavior-preserving.
-63826
+# Co-locate autonomy cascade candidate generators (resident_waits_for_player_gift,
+# resident_mutual_trade/gift/delivery_candidate) from main.rs into autonomy.rs
+# (was resident_offer_scoring.rs). Rename module to reflect its role as the
+# autonomy decision system home. Delete dead RippleBudget zone branch where
+# frontier and sanctuary returned identical budgets. Behavior-preserving.
+63585
